+++ packages/dbsync/CHANGELOG.md
@@ -6,6 +6,10 @@ While in pre-release, assume that any change is a breaking change until v1.0.0 i
 
 ## UNRELEASED
 
+## 0.0.64
+
+## 0.0.63
+
 ## 0.0.62
 
 ## 0.0.61